requirements.txt
niquests
luma.core
flask
